Remove dependency on poise-python
Poise-python was listed as a dependency in metadata.rb for the gigadb
cookbook but is not present in the chef-cookbooks repository. Since
the python cookbook is already listed as a dependency in metadata.rb
to provide python functionality and poise-python is not included in the
chef-cookbooks repository, its dependency has been removed.

The problem is that the python cookbook is deprecated. It is suggested
to move to poise-python (https://github.com/poise/poise-python).