Update .Vagrantfile to use Ubuntu-14.04 box
A problem occurs when booting up nginx during chef provisioning. An
error message, “invalid value "TLSv1.1" in /etc/nginx/nginx.conf:32”
is generated by the lucid32 machine which is running Ubuntu-10.4.
This version of Ubuntu is using nginx version 0.7.65.

The change to Ubuntu-14.04 (Trusty Tahr) fixes this problem. This
version of Ubuntu is using nginx version 1.8.1.