CasperJS test for admin interface: navigation to submission wizard
After testing navigation to the profile page, we test navigation to
submission wizard.  this test need future improvement as the
interaction with the “Submit new dataset” button on the profile page is
not tested (we load the next page directly) as CasperJS is throwing
unresolved errors when attempting to do so.  This work package is to
update the form fields on the submission wizard, so that interaction is
not on critical path so improvement can wait for a later time.