Fixed #70
Re-arrange layout of the form used by admin user to update a dataset
metadata, so that the title and description occupy a prominent place at
the bottom of the form while the rest of the fields are still laid out
in two columns above the title and description.

This change doesn’t include the keyword and Mint DOI fields and button
as they are covered by their respective ticket.