Implementation for #69 (keywords field to admin dataset form)
There is now a Keywords field on the dataset admin page.

Keywords are entered in the text field one by one separated by commas
or by pressing return key.

Keywords can be removed by clicking on the (x) symbol next to it in the
Keywords field.

Upon dataset update, old keywords are removed from the database and the
new set from the form is created in the database as Dataset Attributes.

The keywords are displayed on the dataset view just under the
description.

TODO: automated UI tests, behaviour when JS is turned-off.