2/2 of #41: redirection interestitial
when a dataset page as a dataset attribute with a url to redirect, when
navigating to and only to that url, the user is shown an interstitial
with a redirect notice before being redirected 5 seconds later to the
DOI url.

the metaData array in the Controller parent class has been amended with
a ‘redirect’ attribute whose value is being set in the Dataset
Controller and tested in the  dataset layout to decide wether or not to
show the interstitial.

The interstitial HTML headers has a meta refresh redirect set to 5s
pause and to the DOI as redirect url.