Dataset model: added toXML() public method
Added a toXML() public method to the Dataset model that serialise a
dataset into a Datacite XML schema 4.0.

Used the example XML file from #37 as a starting point.

This new public method is to be used in the Mint action from the
Dataset Controller in order to implement  #38.

the schema is not yet complete as they are some information in the
example XML file that I don’t know where in the database it comes from.
(notably the available date and the related identifiers).