tweaked form alignment
fixed form alignment for the right hand side column of the form so all
text input right-align.

The jQuery tag-editor overlay its own large input on top fo the form
input for keywords, so all text input from the right-hand side column
needed to be enlarged to the right to match that javascript created
input.