Refactored file table (for #63)
added an anchor around file table to allow visitor to jump directly to the file table without scrolling.
It will be especially useful later when the ftp table and navigation between ftp table and file table are implemented.

created two partial views, one for file table, one for ftp table

updated Dataset Controller to accept location parameter and pass a flag to the view rendering so the view know if it needs to render the file table or the ftp table