testing connectivity between web site and ftp server (for #63)
For development, both the web site and the ftp server needs to be on the same network in order for communication to happen,
so i modified the vagrant network configuration for the gigadb-website to specify same private network as ftp server.
On production, both servers will be already on the sames public network (Internet).