Initial implementation of ftp table (for #63)
On the file table, directories are clickable and upon click, an ftp table with the directory listing is loaded.

The ftp table table display directory listing and sub-directories can be traversed while clicking on a regular file will trigger a dowload from ftp server.

Pagination and bread crumbs are NOT yet implemented for the ftp table.

the File model was updated to show the url for loading the Ftp table for directory entries.
a DirectoryListing component was added as a data object for a directory extending on the GFtpFile component to allow additional metadata for self-reflexion (am I a directory?) and full download url.