implementation of custom pagination for ftp table (for #63)
Created a custom pagination for the ftp table so to preserve the location parameter that determine which directory the ftp table is for.

To that effect, I've moved the Array Data Provider from teh grid widget partial to the Dataset Controller, so that pagination can be customised.

Created a new component FtpTablePagination that extends CPagination and allow generating paging urls that preserve locaion paramters and the file table anchor.

Removed some debug codes