worflow and infrastructure for #67:
For the multiple download functionality
it is necessary to handle the packaging of the files as asynchronous jobs
performed by workers scripts because of the processing time for creating
the archive and the large size of the files.

After comparing multiple options, We will use beanstalkd [1] as a job queuing management system and supervisord
for managing the worker processes. PECL Archive_Tar will be used to prepare the
downloadable archives.

This commit contain partial integration of beanstalkd with the gigadb-website for
submitting file packaging jobs when a user selects multiple files for download.

An ajax button is added to the file tables on dataset view to allow users to prepare
a downloadable bundle of their selections of files.

An initial version of a worker script implemented as a Yii Console command
has been started with ability to pull tasks from a beanstalkd job queue
and preliminary integration with Archive_Tar (wip).

Both the yii web app config and console config needed modification too.

TODO: finish worker (integration with FTP, archive building,
job housekeeping, disk space housekeeping, supervisord integration)

[1] https://github.com/kr/beanstalkd/blob/master/doc/protocol.txt