uploading the package of selected files for #67
started work on another worker (not yet functional) to upload the tarball of selected files to Amazon S3.
The file bundling worker will post a new job into a work queue upon bundling success for the uploading worker to consume.
The AWS SDK for PHP v2.8.28 is included in the vendors/ directory.
Because the tarball can be very large we use the S3 MultipartUpload feature in the uploading worker.
We use separate workers for bundling and uploading because they stress different resources (cpu, disk space & download bandwith vs upload bandwith)