upload tarball of file selection to S3 for #67
Once the files selected by a user have been downloaded and archived by the bundle worker.
A new mesage is sent to the job queues systems picked up by the upload worker which upload the tarbal to AWS S3.
the AWS access_key and secret_key  for console based upload worker are managed as configuration similarly to the config for Database and Elastic Search.
Added an AwsYiiConfig application component to make it easier to integrate the use of AWS features (S3 here) with gigadb website.

Updated Chef provisioning to generate a config for aws credentials.