bug fix for #67
remove error handling code that wasn't catching any of the commonly occuring errors.
fixed an issue where the worker was unable to reserve the next job after the first job has been reserved and processed
(because of redundant call to $consumer->connect() in the function to post a new job for uploading the bundle that reset the connection for the bundle job processing.