link to allow visitors to download their selection of files #67
after confirming selection, visitors can click on a page that will show a link to the files bundle tarball when it's ready and uploaded to S3.
If the bundle is not uploaded yet, they will see a message inviting to refresh at a later time.

tweaked the AWS configuration component to work both for console commands and webapp controllers.