Enabled recursive download of directories fro #67
Fixed varsious issues with paths in the tarball.
Fixed issue where the dataset_id wasn't passed to the bundling worker.
Fixed issues where regular files on the dataset table  (therefore directly at the top level of a specific dataset directory) were not handled properly
Enabled checkbox next to directory in the dataset file table that allows download of the content of that directory