started implementing workflow for file previewing (#66)
nothing functional yet.
ajax preview button on the ftp table will either open an iframe with the preview or trigger the submission of a preview generation job
based on wether there is a preview url for the file location stored on Redis.
Created a new Preview action for the FileController to handle the ajax request. Created A new Yii console that will generated the preview
as a background task after downloading the file from the FTP server. The same worker will upload the generated preview to S3 too.