Implementing more error scenarios for previewing system (#66)
If issues with ftp connection, during initial connection or download,
we release the job so we can try later when connection is more reliable.
Also if the file is a binary file and has been partially downloaded,
we can resume download at the last position when a job is being retried.
Added job id in logging message when available.
Added FASTA fa file extension support in the ftp mode detection (binary/ascii).
Cleaned up some variables.