Refactored the file format detection to avoid downloading files that are not supported yet for preview. (for #66)
That include detection of mime type just based on the extension of the file rather that mime magic analysis of its content.
It means we have to maintain a hand-crafted list of extensin/mediatypes pairs.
That allows us to add support for bioinformatics format that are not supported by OS mime magic files.