integration of MFR to the preview functionality (for #66)
Added an yii application component to allow configuring integration of a MFR server.
Construction of the preview url now checks whether an MFR server is running so it only use MFR if there's one running.
Otherwise the preview url is displayed directly in the preview modal iframe.
Added ability to configure the MFR component with the extension supported by MFR.