Configuration and availability of preview service (for #66)
now the mime types that our preview code support as well as the file extensions that MFR supports are now configured in one json file read by web site and background workers..
Because that json configuration file is not environment dependent it is created into the config/ directory with no need of chef provisioning involvment.
The status of the MFR preview server is checked whenever the preview button is clicked irrespective of the status of the preview process for that file.
That way if the MFR was shut down while a visitor click on the proview button  and it is then brought online again,
the visitor will get the MFR experience if he/she clicks the preview button again