Work to get background worker to run on ftp server (#67)
Replaced the upload to s3 of file bundles by a local copy into the FTP server root.
Replaced the s3 object check by Redis based status.
Tweaked the Redis conf to be less secure as this is dev environment.
Removed Supervisord from the queues server as redis and beanstalkd can run as system managed daemon
and dont need the added value of Supervisord