Affilate Login (#64): acceptance tests  Setup the BDD framework (behat) to automate the acceptance tests.
no need of phantomjs for now, so removed it from the doc.
Finally figured out how to use PHPUnit assertions inside the Behat FeatureContext file.
Tweaked the scenarios to make the naviaton action more flexible.
Added comments that the context  in the  first scenario is set to pass so i can test that interaction with browser
is set up correctly. Will revert them to pending prior to start test-driving the implementation of the feature.