Affiliate Login (#64): progress with setting up the BDD framework
Implemented the functional tests that check that Gigadb has registered api keys configured for all the OAuth providers.
(Currently failing has i have only registered with Twitter OAuth API so far)
Those tests are the steps for the new background scenario in the user story's acceptance tests for this feature.
Also tweaked the formatting for the scenarios.
Installed wunit for functional testing in Yii without selenium.
Also the acceptance tests now (may not be definitive) need to be run in the virtual machine so it can find the Yii framework.