Affiliate login (#64): Sign in with Twitter done
Sign in  with twitter working according to specs.
Instead of changing findAffiliateUser static method that could be use elsewhere in the codebase, I've added a findAffiliateEmail just for the purpose of find an affiliate user by email to prevent a user having multiple account for the same email based on how they log in.
Tweaked the acceptance tests to be able to create a new user account to test the above criteria.