Affiliate Login (#64): working on sign in with Facebook
started work on Facebook signin.
Updated the scenario to be consistent with the Twitter scenario and in order to reuse the same steps of functional testing.
Updated the FeatureContext steps to be more generic and work across provider (at least Twitter and Facebook).

Curently stuck on the interstitial page after being authorised and before redirect to the gigadb web site: behat/mink are unable to find and click the submit button on that page for some mysterious reasons.