Affiliate Login (#64): Update User record whenever affiliate login is used
Even though user has already a gigadb account, and even though any existing account is searched using email, when a user log in using the affiliate provider of their choice, we still need to update the user record with the provider-supplied uid  so that session data matches user record and session code does not need to change. Also make user record consistent and information-rich (we know all of the provdiders a user is using to log in).

Updated the second Facebook scenario to re-use existing steps definition.
When all facebook scenarios are run together, second one is failing because I still haven't figured out to deauthorise programmatically the test-user from gigadb facebook app. I may have found a solution in using the Graph API to create/delete test users but it is yet to be implemented (todo) in the steps definition.