Affiliate Login (#64): fixed interactions between automated tests and database
When inserting data into a pgsql table with autincrement, I need to use nextval() on the autoincrement sequence.

The root cause of the duplicate key error, though is that the database fails to be cleared between scenario sometimes with the error: ERROR:  database "gigadb" is being accessed by other users.

Fixed it by calling pg_terminate_backend on all backend processes connected to gigadb database before dropping the database.