Affiliate login (#64): login with ORCID
Implemented a revoke endpoint to allow a user of the current session to revoke gigadb authorisation for his/her oauth provider directly from gigadb web site. Also useful for clearing state in automated tests. Only work for ORCID for now.
The revocation is idempotent and for unsupported provider it justs logout the user.

All implemented test scenarios (with the @ok tags) are now passing.