Auhtor-names(#81): step definitions for the acceptance tests for gigascience story
implemented the step definitions for the scenario to implement the story for issue #81 from gigascience perspective.

Made used of Behat tables to and real author names from the database.

That means taht before each run of the suite, the sample of production database needs to be loaded. The database dump is called "author-names-80-81-82.pgdmp" and it differs only from the  2016 dump by having an Attribute record for "urlredirect".