Author-names (#81, #82): set up unit test framework
It was tricky to get PHPUnit with Yii1.1 with Gigadb-website running in Vagrant.
Among the several problem:   recent version of Php unit and its extensions would cause issues (4.8 fails. but 4.1 is ok).
Also the phpunit selenium extension is required even we don't use it  because the Yii framework references it!
Also avoid using minimum-stability as 'dev' in composer.json. prefer 'stable' and override with @dev on individual package if necessary.