Accentuated characters not displaying correctly (#82): fix in progress
work on fix to  #82 by replacing php string operations by unicode aware functions from mbstring.

The mbstring php extension is not installed by default, so needs to update gigadb provisioning to include the Chef recipe for mbstring.

The problem is that it seems the php.ini file needs to be updated with some minimal configuration parameters for mbstring, but the Chef recipe for php::module_mbstring is not doing it:

[mbstring]
mbstring.language = Neutral
mbstring.internal_encoding    = UTF-8
mbstring.encoding_translation = On

So I still needs to figure either how to get the provisioning mechanims to update php.ini or to make the configuration change unecessary.

I've also improved unit tests coverage for this functionality to cover Dataset methods that are related to authors' names.