Accentuated characters not displaying correctly (#82): fix complete
second part of the fix to #82.
Added  a php configuration setting to enable the use of multibyte unicode aware string functions from the php module mbstring.

That setting is only needed when running PHP <5.6 as from 5.6, that setting is deprecated and the  default setting for characterset is UTF-8.

Updated both the main config and the test config as the unit tests are running from the test config.