Affiliate login (#64): fix to a fix
During dev, I made a change to method signature for generatePassword, in order to fix a bug introduced in previous implementation of this feature, not realising that the method is used elsewhere. Reviewing the change, I realise that method signaure change was unnecessary to fix the initial issue, so reverted that change and fix the original issue by using the method correctly as it's done in other classes.