author-names (#80, #81, #82): fixing the tests
downgraded PHPUnit to 4.1.1, so it work with the Yii Framework (even 4.3.* doesn't work with Yii 1.1).
Switched to minimum-stability: stable, so I can remove the @stable suffixes
add @dev suffix to mink-wunit-driver as that's the only branch avaiable for that package.
Removed google api client as it's already in Gigadb codebase.

Because of the PHPUNit downgraded, cannot use namespace when using PHPUNit Assert
(PHPUNit\Framework\Assert had to be replaced by PHPUnit_Framework_Assert)

Controlled which Behat hooks get run with which scenario using tags on features and on hooks
so that for example oauth revokation hook are only actioned in affilate login scenario

Addded a scenario to ensure the test environment is loaded up with all the test data in the author name display and name display preview tests.

Updated one sql test data script to drop the not null constraint on the affiliation column of the gigadb_user table.

Updated the test runner so it run all the acceptance tests and unit tests.