author-names (#80,#81,#82): make the loading order of database fixture deterministic
By default CDbFixtureManager.php loads the fixtures using readdir() which returns list files in the order in which they are stored by the filesystem. It "seems" than that order vary from system to system resulting in the following error:

CDbException: CDbCommand failed to execute the SQL statement: SQLSTATE[23503]: Foreign key violation: 7 ERROR:  insert or update on table "dataset_author" violates foreign key constraint "dataset_author_author_id_fkey"
DETAIL:  Key (author_id)=(1) is not present in table "author".

because the join table fixture is loaded before the data table.

I've created an init.php file in the fixtures directory to customise the fixture loading behaviour, in this case, ensuring the fixtures are loaded  in the order that won't violate the foreign key constraint.