display datasets on user profile (#60): load a user in acceptance tests using sql
In order to tests that a user can have datasets of author is attached to displayed on his/her profile,
we make it possible to create a new user that can sign in from the scenarios.

This is because the default users in test data already have the same tests datasets listed on their profile because
they submitted them.

we want a user with no existing link to datasets prior to running the tests