user claims dataset (#57): catering for edge cases for user claims
added a scenarios and implementations for canceling claim
and preventing a user with pending claim to submit another claim

That includes added static function in Author to find and return an author instance that is attached to the supplied user id
And fixing the javascript used to control the visibility of the summary message to the user.
Tweaked the error messages and make sure they are generated in the controller.