admins attach user to author (#56): adding missing edge cases, simplified workflow
Refactored workflow for coherence/clarity and to cater for edge case from integrating with #57

added two actions in AdminAuthorController.php to deal with the two stages of linking a user to an author
when starting from user administration screens (user view or user list):
actionPrepareUserLink set and unset the session variable to store user_id when the admin user is on the admin author list to pick an author to link to
(having already selected a user that then need to be stored in a session variable)

In adminAuthor/admin.php: Added various ways to unset the variable when it's not needed:
clicking on the "Abort and clear select user" or by clicking the close button of the alert messages
Also set up a modal dialog with a link and abort buttons. Dialog is opened when the admin user click on an author row and the author id is passed on to the dialog

In  user/admin.php:
fixed the link to go to the refactored controller action for selecting authors to link to.

various UX tweaks on other files.

Removed some redundant code form the tests scenario and from the adminAuthor/form.php due to change in workflow