in the process of implementing authors merging (#49), I have added an unMerge method
the unMerge method will remove an author from the graph by removing all the outward edges from itself.

I also stumbled into the following errors when writing tests that create or delete an AuthorRel object in database:

CDbException: CDbCommand failed to execute the SQL statement: SQLSTATE[23505]: Unique violation: 7 ERROR:  duplicate key value violates unique constraint "author_rel_pkey"
DETAIL:  Key (id)=(1) already exists.

This is because the data fixtures for author_rel are loaded before the tests are run and for some reason either CDbTestCase or CdbFixtureManager set db connection are unable to auto-increment from the last id used in the fixture.

I found a workaround from this issue (which is for a different context/framework version but same problem):
https://github.com/yiisoft/yii2/issues/8298

Didn't put it in a setUp function as it makes the other tests fail.