Creating the dialog box for merging two authors (#49)
Created a Bootstrap modal (whose dimension can be better controlled than Jquery-UI/CJuiDialog dialog) for the pop up box
Displayed the details of the two author to merge in the pop up was tricky as the details for the the author to merge with
is only known by Javascript, php doesn't have access to it yet, while the details for the author we want to merge come from php
I used the 'rowHtmlOptionsExpression' of CGridView to inline a an author's details as data-attributes to the <tr>
When a row is clicked, the open_control javascript that manage the dialog copies the data attributes to the comparison table.

To add the server side functionality fo merging the authors, I added an mergeAuthors action in the AdminAuthorController
that calls a merging method on the Author model class.

The base url is passed on to Javascript code using registerScript at the top of the admin page. It's cleaner than having all javascript
functions to do their own thing. At the moment only merge_authors() javascript functions use it. but will refactor the other function later.

upon merging success, a flash message is displayed on the admin author view.