in the dialog box for merging two authors (#49)
in the comparison table showing the details of the two authors, the admin can now see the graphs of
identical authors they each has already been merged with.

This required a new action in adminAuthor controller to return through ajax request the display names of the graph for the target author if it exists.