Enable GitLab secrets in config and test scripts, streamline variables
Additionally, update env-sample with all the necessary variables (no more, no less),
with appropriate defaults.

The default for DATA_SAVE_PATH is now in user's home directory to avoid data being wiped out
if the local git repo is removed.

the PostgreSQL data is now back to be bind-mounted for consistency
(all project data are now in DATA_SAVE_PATH) and because
the benefits of using an explicit volume (and potential volume drivers)
are important only on staging/production env which will have its own compose file
override anyway, and we may not do persistence on first iteration anyway.

Removing COMPOSE_PROJECT_NAME, as it doesn't add value
(when using docker-compose CLI, service names are used, not containers'),
but the removal had ramifications (e.g: broke restart_php.sh which broke tests)
but found alternative comforting me that that variable wasn't necessary
and over-used across too many areas of the codebase.

The generate_config.sh script is now better isolated for running in-container
(previously it was dependent on host-context variables like:
 DATA_SAVE_PATH, COMPOSE_PROJECT_NAME, APPLICATION)

Fixed a directory manipulation issue with downloading the YII framework
onto a bind-mounted path (YII_PATH).

To enable the tests to access GitLab secrets written in the .secrets file,
they need to be exported to the environment
(as they are accessed using PHP's getenv).
To do so, the sourcing of the file needs to be wrapped with bash's 'set -a'
and 'set +a' which respectively enable and disable auto-export to the environment.