Refactor get curator name to separate responsibility between dataset/user
In the dataset admin view (dataset/admin.php), in the CGridView widget,  there is no need to use $data->curator_id
and passing it to a static method on Dataset to retrieve the curator's name.

It works, is more maintainable and is simpler to just call '$data->getCuratorName()'

However forming a full name for a user is the responsibility
of the User class, not of Dataset.

So the getCuratorName() in Dataset class should delegate to a getFullName()
method on a User class.

We have now unit tests for the described behaviour.

As a side note, those tests increase the coverage from 27.16%
(which is down from 27.19% after merging the work enabling curation_log)
to 27.34%