Get Mink accessible with subcontext for easier migration to Behat3
Third part of making changes to accommodate migrating from Behat 2 to Behat 3.

in particular Behat 3 doesn't support communication between context using subcontexts.

So to make it easy to transition Behat environment based communication I need to make everything subcontext based
including the main context and the access to Mink methods, so that I can migrate with the single search and replace.

This commit is about to make Mink methods accessible through subcontext, previously the main context, a subcontext like all the others.
In its place, I create an empty FeatureContext as the main context.

see: https://stackoverflow.com/questions/29674685/how-to-communicate-between-contexts-in-behat-3