Set byte conversion library, update file size calculation with formatting
I had to choose between metric or binary system for file size basis.

Went for metric as that's what used for storage media (hard drives, DVDs)
and networking, and on macos
and also it looks nicer.

See: https://en.wikipedia.org/wiki/Gigabyte

Note1: the previous code used binary system for calculation and value display
but it displayed the wrong units

Note2: it's easy to switch between the two systems, but make sure to keep the tests updated
(both unit and functional)