Add Service for replacing keywords in database after sanitizing input
GitLab SAST report found three high vulnerabilities in DatasetController.

It's actually the same code snippet duplicated in three places in that class.

The snippet's function is to delete keywords in database and replace them
with keywords string from POST data.

So we create an AttributeService that will:
- encapsulate that function in one place to remove duplication
- remove the unsafe code and sanitize input
- stop DatasetController from dealing with storage concerns (that's not what controllers are for)

To that end, I had to fix the way database fixtures work:
 - Removed the init.php file as it was in the way of database constraints
 - Instead added special initialisation sequence for dataset and attribute fixtures (see dataset.init.php and attribute.init.php)
 - Updated the RelationDAOTest to not require a pre-test reset  as datasets are only used for reference data in this test

Also created a DatasetDAO to manage Dataset and Dataset related persistence logic (currently, just the DatasetAttributes).

I had to use Dependency Injection and to create a DatasetAttributes factory class to enable TDD and clean design of the DAO layer.

Updated composer.json.dist so all the new classes can be found and used anywhere in the codebase.

The work is not finished yet, TODOs are updating the controllers to use the new service, test-driven with functional tests.

This is part of the work for ticket: https://github.com/rija/gigadb-website/issues/121
More specifically it covers the "Fix unsafe code" task (as well as falling under the "Remove duplicate code" task as a side effect)

It also lays the foundation for the future reworking of DatasetController.