Use AttributeService in create1 and datasetManagement actions
in order to remove unsafe and duplcated code as part of ticket:
https://github.com/rija/gigadb-website/issues/121

I had to update keyword replacement method in AttributeService
to detect whether it is being called inside a database transaction already.
As the DatasetController actions create1 and datasetManagement already
initiate a transaction while the update action (first to use the new AttributeService)
doesn't.

Also I had to fix a bunch of Javascript errors on _form1.