Upgrade Google API client to latest with modern json-based credentials
The use of "Google_Auth_AssertionCredentials"
(that require a binary P12 private key) is deprecated

We now authenticate to the API the recommended way
using the service account json private key.

Added a functional test to verify the google client authenticate correctly.

In order for the test to run on. CI, I have to store the json private key
in GitLab secret variables (on project level) and modify config_generate.sh script
to generate the file dynamically.

in generate_config.sh, we need to process the ANALYTICS_PRIVATE_KEY variable
separately from the other secret variables because it's multiline with special characters
and it would break when sourcing all the shell variables.

That's also why in the .gitlab-ci.yml config, we also exclude it form the .secrets generation