Replace Efeed with Suin\RSSWriter using Composer, fix the RSS feed
as part of the work to move all dependencies to Composer:

https://github.com/rija/gigadb-website/issues/123

I've replaced old Efeed library with an RSS writer library from Packagist.org

Also it turns out that the RSS feed in GigaDB were actually not implemented:

 /site/feed returned a valid RSS feed for a dummy example.

I fixed that.

the new RSS library is modern and has 100% test coverage.
However it doesn't seem to be actively maintained.

I've also created a new service (NewsAndFeedsService) to manage all the features around the RssMessage and News tables
as well as the latest dataset to be notified in the homepage and in the RSS feed.