Manage lessphp with composer, set vendor autoload for console commands
because when the webapp start we need to run both composer update
and then compile the less files, I created a new ops scripts to do both.

the version of leafo/lessphp is set to 0.3.9 only because newer versions
break are breaking API compatibility (BC)with the version of bootstrap
still used in the admin side of the app and depended on by the site.less file.