Setup Opauth with Composer, fix affiliate login acceptance tests
As part of work on https://github.com/rija/gigadb-website/issues/123

to have all PHP dependencies managed by Composer, I have removed the files
for the Opauth library and its Services specific Strategies from the codebase.

Instead i added the required libraries to the composer.json.dist file.

version of the code is slightly different so main.php.dist had to be changed
to reflect the proper way of configuring Opauth.

Improved security by defining a stronger security salt for Opauth
and storing it as a GitLab secret variables.

Removed the long redundant actionChooseLogin in SiteController
and the corresponding view.