Replace Mailchimp client with new one installed with Composer
This is part of the work to manage all PHP dependencies with Composer:
https://github.com/rija/gigadb-website/issues/123

Also create a new service (NewsletterService.php) to manage the list
(subscribe and unsubscribe) instead of using the grab-all Utils.php class.

The NewsletterService.php is an Yii Application Component so it is configured
in the local.php.dist.

The functional tests for the service requires the developer to set up
a Mailchimp account and and email service that setup with a catch-all email address.
Although Mailinator.com fits that constraint, it cannot be used as it is banned on Mailchimp.
The info about the dev account should be in the following variables in the GitLab secret variables service
for the developer's fork:
MAILCHIMP_API_KEY
MAILCHIMP_LIST_ID
MAILCHIMP_TEST_EMAIL