Abstract browser mediation in functional tests so it's easier to change
as part of work to migrate to PHP 7.0, I found issue with phpunit-mink.
It is known by their developers since 2016 and the project hasn't had activity since then.
Before I can change to a more modern browser automation framework,
I need to coral as much as possible all the browser automation into specific abstractions.

In this case I get the FunctionalTesting that contains phpunit-mink dependency
and common setup and tear down operations to be the parent for all my functional tests
and the Browsers*Steps.php traits will contain all the code related to interacting with the browser

Also created a CommonDataProviders trait for common data providers used by several tests.

Frustratingly, the traits needs to be required once in bootstrap.php
despite adding a PSR0 autoloader for them in composer.json.dist