Defer javascript execution to improve page rendering
First had to fix existing javascript errors, ensuring that we are loading
recent versions of jquery, jquery-ui and bootstrap only once on each page.

Then added the "defer" to the <script> tag for external libraries so they are executed
only when the DOM is fully loaded, including jQuery, so to avoid blocking time during loading of resources.

see:
- https://developer.mozilla.org/en-US/docs/Web/HTML/Element/script#attr-defer
- https://html.spec.whatwg.org/multipage/scripting.html#the-script-element

That means that all javascript code that depend on jQuery on pages need to be surrounded by

document.addEventListener("DOMContentLoaded", function(event) {
...
});
otherwise they will error because jQuery is not loaded yet as the script is being parsed and executed by the browser.
The DOMContentLoaded event is fired after all the deferred scripts are executed.

That works fine for code in <script>, but needed:
(1) to unwrap script wrapped in clientScript->registerScript() because that function insists in writing automatically Jquery to the page
(2) ajaxButton and ajaxLink to be deconstructed into a Bootstrap button and a standalone inline javascript function for same reasons as (1)
(3) to subclass CJuiAutoComplete Yii widget so that its run() method don't use registerScript()

Those effort have yielded significant reduction in loading time on the home page, the search result page and most of all on dataset page