Build nginx container and only share assets between web & app containers
In order to deploy to a production environment , both nginx and php-fpm
containers need to be fully configured and as immutable as possible.
Also we may have limited access to the deployment host (no ssh),
so as much of the config and application files need to be baked in containers.

Also nginx need to be able to read the static files as well as the  asset folder
while the yii application needs to write in the assets folder,
so assets need to be a shared volume.

Because all of that the official Nginx container cannot be used as is,
instead we derive a new image from it (see Web-Dockerfile and docker-compose.build.yml)