Make coverage check more robust
by default, it use GitLab API to provide reference coverage value.

If the token is not set (e.g:  a contractor who didn't set up a gitlab account or an opensource contributor), it reverts to Coveralls.io

If no access to coverage info about main branch, it reverts to a default value.