Implements test scenarios for the samples tab on dataset view
It similar implementation as for the already implemented files tab scenarios.
So i could generalise the step definitions to works for both tabs.

However I run into the issues that the UI language on the table settings
is the same for both tabs, which is good for the user.

But it caused the tests for the files tab to fails, because the code for both are in the same html page
, with javascript used for routing to one or the other.
the Mink traversal functions are triggered by the the first one they encounter with the form elmeent name
(in this case the sample form elements), causing test failure with the following errors:

"Element is not currently visible and may not be manipulated"

the solution, which also happens to be good practice is to ensure all form elements have and "id" attribute
and that this id is unique. Those will be the target of the step definitions.